[major] Layer-associated probes semantics and ABI
Define semantics and an ABI for layer-associated probes.  This introduces
a major change to the existing layer ABI where layers are now enabled with
a Verilog text macro (`define) and not by including a file.  This is done
to unify the enabling of all layer collateral which can include
modules/logic _and_ new probe ports (which were not supported previously).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>